+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased][unreleased]
-- 
+-
+
+## [1.1.2] - 2016-02-01
+- Added an gateway settings class.
+- Added support for creditcard payment method.
 
 ## [1.1.1] - 2015-08-04
 - Use wp-pay/core XML security filter function.
@@ -26,7 +30,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2014-12-12
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/sisow/compare/1.1.1...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/sisow/compare/1.1.2...HEAD
+[1.1.2]: https://github.com/wp-pay-gateways/sisow/compare/1.1.1...1.1.2
 [1.1.1]: https://github.com/wp-pay-gateways/sisow/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/wp-pay-gateways/sisow/compare/1.0.3...1.1.0
 [1.0.3]: https://github.com/wp-pay-gateways/sisow/compare/1.0.2...1.0.3